Harmonize header definition of stress_misc_stats_set

+++ stress-ng.h
@@ -2501,7 +2501,7 @@ extern WARN_UNUSED int32_t stress_set_vmstat(const char *const str);
 extern WARN_UNUSED int32_t stress_set_thermalstat(const char *const str);
 extern WARN_UNUSED int32_t stress_set_iostat(const char *const str);
 extern void stress_misc_stats_set(stress_misc_stats_t *misc_stats,
-	const int idx, const char *description, const double value);
+	const size_t idx, const char *description, const double value);
 extern WARN_UNUSED int stress_tty_width(void);
 extern WARN_UNUSED size_t stress_get_extents(const int fd);
 extern WARN_UNUSED bool stress_redo_fork(const int err);
